+++ src/main/java/de/hhu/propra/sharingplatform/controller/ItemController.java
@@ -1,6 +1,5 @@
 package de.hhu.propra.sharingplatform.controller;
 
-import de.hhu.propra.sharingplatform.dao.ImageService;
 import de.hhu.propra.sharingplatform.model.Item;
 import de.hhu.propra.sharingplatform.model.User;
 import de.hhu.propra.sharingplatform.service.ItemService;
