+++ src/main/java/de/hhu/propra/sharingplatform/service/ImageService.java
@@ -1,4 +1,4 @@
-package de.hhu.propra.sharingplatform.dao;
+package de.hhu.propra.sharingplatform.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
